Resolve warnings generated with compiler option -Dconversion
Most of the warnings were resolved by changing the data types of some
variables. Others required casting, particularly where kernel interfaces
are involved.

There were a few instances discovered that were errors, for example comparing
an unsigned int against -1, and assigning a 16 bit value to a uint8_t.

This commit also adds configure options --enable-conversion-checks and
--enable-force-conversion-checks, the former adds compiler option -Dconversion
unless the compiler is an old version that throws up false warnings. Option
--enable-force-conversion-checks adds -Dconversion even if the compiler throws
up known false warnings.

+++ keepalived/core/process.c
@@ -35,7 +35,7 @@ set_process_dont_swap(size_t stack_reserve)
 {
 	/* Ensure stack pages allocated */
 	if (stack_reserve) {
-		int pagesize = sysconf(_SC_PAGESIZE);
+		size_t pagesize = (size_t)sysconf(_SC_PAGESIZE);
 		char stack[stack_reserve];
 		size_t i;
